requirements.txt
# ================================
# Core ML & Scientific Computing
# ================================
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.2
torch>=2.0.0

# ================================
# Data Collection & Processing
# ================================
yfinance>=0.2.28
requests>=2.31.0
beautifulsoup4>=4.12.0

# NLP / Sentiment (optional, auto-fallback supported)
transformers>=4.36.0
sentencepiece>=0.1.99

# ================================
# Visualization
# ================================
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# ================================
# Dashboard / Web
# ================================
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# ================================
# Utilities
# ================================
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
joblib>=1.3.2

# ================================
# Testing
# ================================
pytest>=7.4.4

# ================================
# Notes
# ================================
# - Python version: 3.11 (recommended)
# - sqlite3 is part of Python standard library
# - Conda is recommended for installation to avoid ABI issues
